Fix CDC Building Class pie chart rendering and data type issues
- Convert COUNT(*) to INTEGER in SQL query to prevent bigint string conversion
- Ensure all total_count values are converted to Number in chart data
- Increase chart size (200-350px) and pie radius (70-130px) for better visibility
- Fix percentage calculations in table with proper number conversion
- Resolves issue with total count showing as concatenated string
- Resolves issue with blank pie chart area

+++ src/app/api/cdc-comprehensive/route.ts
@@ -191,7 +191,7 @@ function buildPieChartQuery(lgaCode?: string, lgaName?: string, startDate?: stri
   return `
     SELECT
       building_code_class as building_class,
-      COUNT(*) as total_count
+      COUNT(*)::INTEGER as total_count
     FROM housing_dashboard.cdc_records_raw
     ${whereFilter}
     GROUP BY building_code_class
